# vision-service/openapi.json
{
  "components": {
    "schemas": {
      "Body_calibrate_v1_calibrate_post": {
        "properties": {
          "frames": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Frames",
            "type": "array"
          }
        },
        "required": [
          "frames"
        ],
        "title": "Body_calibrate_v1_calibrate_post",
        "type": "object"
      },
      "Body_validate_v1_validate_post": {
        "properties": {
          "calibrationToken": {
            "minLength": 16,
            "title": "Calibrationtoken",
            "type": "string"
          },
          "frames": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Frames",
            "type": "array"
          },
          "sequenceNo": {
            "minimum": 1.0,
            "title": "Sequenceno",
            "type": "integer"
          },
          "turnId": {
            "title": "Turnid",
            "type": "string"
          }
        },
        "required": [
          "turnId",
          "sequenceNo",
          "calibrationToken",
          "frames"
        ],
        "title": "Body_validate_v1_validate_post",
        "type": "object"
      },
      "Box": {
        "properties": {
          "h": {
            "exclusiveMinimum": 0.0,
            "maximum": 1.0,
            "title": "H",
            "type": "number"
          },
          "w": {
            "exclusiveMinimum": 0.0,
            "maximum": 1.0,
            "title": "W",
            "type": "number"
          },
          "x": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X",
            "type": "number"
          },
          "y": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y",
          "w",
          "h"
        ],
        "title": "Box",
        "type": "object"
      },
      "CalibrationResponse": {
        "properties": {
          "backgroundClasses": {
            "items": {
              "type": "string"
            },
            "title": "Backgroundclasses",
            "type": "array"
          },
          "calibrationToken": {
            "title": "Calibrationtoken",
            "type": "string"
          }
        },
        "required": [
          "calibrationToken",
          "backgroundClasses"
        ],
        "title": "CalibrationResponse",
        "type": "object"
      },
      "Detection": {
        "properties": {
          "box": {
            "$ref": "#/components/schemas/Box"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "boolean"
          }
        },
        "required": [
          "label",
          "confidence",
          "box",
          "target"
        ],
        "title": "Detection",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "modelVersion": {
            "title": "Modelversion",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "modelVersion"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VisionOutcome": {
        "properties": {
          "elapsedMs": {
            "minimum": 0.0,
            "title": "Elapsedms",
            "type": "integer"
          },
          "level": {
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Level",
            "type": "integer"
          },
          "points": {
            "minimum": 0.0,
            "title": "Points",
            "type": "integer"
          },
          "streak": {
            "minimum": 0.0,
            "title": "Streak",
            "type": "integer"
          },
          "totalScore": {
            "minimum": 0.0,
            "title": "Totalscore",
            "type": "integer"
          },
          "totalXp": {
            "minimum": 0.0,
            "title": "Totalxp",
            "type": "integer"
          },
          "turnId": {
            "title": "Turnid",
            "type": "string"
          },
          "unlockedAchievementIds": {
            "items": {
              "type": "string"
            },
            "title": "Unlockedachievementids",
            "type": "array"
          },
          "xp": {
            "minimum": 0.0,
            "title": "Xp",
            "type": "integer"
          }
        },
        "required": [
          "turnId",
          "elapsedMs",
          "points",
          "xp",
          "totalScore",
          "totalXp",
          "level",
          "streak",
          "unlockedAchievementIds"
        ],
        "title": "VisionOutcome",
        "type": "object"
      },
      "VisionVerdict": {
        "properties": {
          "decision": {
            "enum": [
              "continue",
              "pass",
              "system_error"
            ],
            "title": "Decision",
            "type": "string"
          },
          "detections": {
            "items": {
              "$ref": "#/components/schemas/Detection"
            },
            "title": "Detections",
            "type": "array"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VisionOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "number"
          }
        },
        "required": [
          "decision",
          "progress",
          "detections"
        ],
        "title": "VisionVerdict",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Camera Quest Vision",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/v1/calibrate": {
      "post": {
        "operationId": "calibrate_v1_calibrate_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_calibrate_v1_calibrate_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calibrate"
      }
    },
    "/v1/validate": {
      "post": {
        "operationId": "validate_v1_validate_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_v1_validate_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisionVerdict"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate"
      }
    },
    "/v1/warmup": {
      "post": {
        "operationId": "warmup_v1_warmup_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Warmup"
      }
    }
  }
}